this is for HAWKULAR-511 but since it won't hurt anything (it isn't used anywhere yet) I'm putting this in now into master. I will be committing something soon in the bus that will actually need this setting
We are going to want to change this setting's value in the future because if a message truly is going cross-kettle over the bus, putting the binary blobs in the local data/ directory won't work (the remote kettle won't be able to see it). We will need to either require people to have a shared filesystem across kettles or we need to set this to a shared remote HTTP/FTP endpoint where binary blob files can be uploaded/downloaded.